完成10/25課堂進度檢查

+++ database/migrations/2024_10_22_031842_create_articles_table.php
@@ -17,7 +17,7 @@ class CreateArticlesTable extends Migration
             $table->id();  // 自動增加的 ID 欄位
             $table->string('bikeway_name')->comment('自行車道名稱');  // 自行車道名稱
             $table->string('districts')->comment('行政區');  // 行政區
-            $table->decimal('length', 5, 2)->comment('長度');  // 長度
+            $table->double('length', )->comment('長度');  // 長度
             $table->text('surrounding_attractions')->comment('周邊景點');  // 周邊景點
             $table->timestamps();  // 追加 created_at 和 updated_at 欄位
         });
